requirements.txt
PyGithub==1.55
python-dotenv==0.20.0
requests==2.28.1